Introduction of the AudioDestinationNode module
AudioDestinationNode is a subclass of AudioNode with no constructor and
no extra method. The only effective difference is the `maxChannelCount`
instance property, which is already covered by the
`audioDestinationNode` type in the `WebAudioAPI` module.

This PR allows for `AudioContext`'s `destination` property to be used
like an `AudioNode`, especially being connected to by other nodes since
this is the only way to output audio to speakers.

For more information, check
[MDN](https://developer.mozilla.org/en-US/docs/Web/API/AudioDestinationNode)